% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_rare.R
\name{alpha_rare_curve}
\alias{alpha_rare_curve}
\title{Plotting rarefracation curve for each group with anova statistics}
\usage{
alpha_rare_curve(alpha_rare, metadata, groupID = "genotype")
}
\arguments{
\item{alpha_rare}{alpha rarefracation matrix, typical output of usearch -alpha_div_rare;}

\item{metadata}{matrix or dataframe, including sampleID and groupID;}

\item{groupID}{column name for groupID.}
}
\value{
ggplot2 object.
}
\description{
Input alpha rare and metadata, and manual set metadata column names.
ggplot2 show lineplot, and standard error
}
\details{
By default, returns grouped curve, sample parameter to be continued
}
\examples{
# Set three parameters: alpha_rare, metadata, groupID
alpha_rare_curve(alpha_rare, metadata, "genotype")
# Set two parameters: alpha_div, metadata, groupID as default "genotype"
alpha_rare_curve(alpha_rare, metadata)
# Set three parameters: alpha_rare, metadata, and groupID using site
alpha_rare_curve(alpha_rare, metadata, "site")
}
\references{
Zhang, J., Zhang, N., Liu, Y.X., Zhang, X., Hu, B., Qin, Y., Xu, H., Wang, H., Guo, X., Qian, J., et al. (2018).
Root microbiota shift in rice correlates with resident time in the field and developmental stage.
Sci China Life Sci 61, DOI: \url{https://doi.org/10.1007/s11427-018-9284-4}
}
\seealso{
alpha_boxplot
}
\author{
Contact: Yong-Xin Liu \email{metagenome@126.com}
}
