% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_cpcoa.R
\name{beta_cpcoa}
\alias{beta_cpcoa}
\title{Plotting beta diversity scatter plot of Constrained PCoA}
\usage{
beta_cpcoa(
  otutab,
  metadata,
  dis = "bray",
  groupID = "genotype",
  ellipse = T,
  label = F
)
}
\arguments{
\item{otutab}{rarefied OTU table, typical output of usearch -otutab_norm or -otutab_rare,}

\item{metadata}{matrix or dataframe, including sampleID and groupID;}

\item{dis}{distance type for caucluate, default "bray", alternative "manhattan, euclidean, canberra, bray, kulczynski, jaccard, gower, altGower, morisita, horn, mountford, raup , binomial, chao, cao or mahalanobis".}

\item{groupID}{column name for groupID.}

\item{ellipse}{stat ellipse, T or F.}

\item{label}{sample name showing, T or F.}
}
\value{
ggplot2 object.
}
\description{
Input rarefied OTU table and metadata, and manual set beta distance type and metadata column names.
ggplot2 show CPCoA with color and stat ellipse.
}
\details{
By default, returns beta CPCoA coordinate
The available diversity indices include the following:
\itemize{
\item{most used indices: bray, manhattan, euclidean, jaccard}
\item{other used indices: canberra, kulczynski, gower, altGower, morisita, horn, mountford, raup , binomial, chao, cao or mahalanobis}
}
}
\examples{
# example data: OTU table, rownames is OTU_xxx, colnames is SampleID
data(otutab)
# example data: metadata or design, include SampleID, genotype and site
data(metadata)
# Set 2 parameters: otu table, metadata, and distance type and groupID using default "bray" and "genotype"
beta_cpcoa(otutab, metadata)
# Set 4 parameters: otu table, metadata, distance type as "jaccard" and groupID as "site"
beta_cpcoa(otutab, metadata, "jaccard", "site")
# You can found a two-dimension seperated groups, by watch out p is significantly?
# Set 6 parameters: otutab, metadata, and groupID as using "site",
beta_cpcoa(otutab, metadata, dis = "bray", groupID = "genotype", ellipse = T, label = T)
}
\references{
Zhang, J., Zhang, N., Liu, Y.X., Zhang, X., Hu, B., Qin, Y., Xu, H., Wang, H., Guo, X., Qian, J., et al. (2018).
Root microbiota shift in rice correlates with resident time in the field and developmental stage.
Sci China Life Sci 61, DOI: \url{https://doi.org/10.1007/s11427-018-9284-4}
}
\seealso{
beta_cpcoa
}
\author{
Contact: Yong-Xin Liu \email{metagenome@126.com}
}
