% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_cpcoa_dis.R
\name{beta_cpcoa_dis}
\alias{beta_cpcoa_dis}
\title{Plotting beta diversity scatter plot of Constrained PCoA}
\usage{
beta_cpcoa_dis(
  distance_mat,
  metadata,
  groupID = "genotype",
  ellipse = T,
  label = F
)
}
\arguments{
\item{metadata}{matrix or dataframe, including sampleID and groupID;}

\item{groupID}{column name for groupID.}

\item{ellipse}{stat ellipse, T or F.}

\item{label}{sample name showing, T or F.}

\item{dis}{distance type for caucluate, default "bray_curtis", alternative "unifrac, unifrac_binary, jaccard、manhatten, euclidean".}
}
\value{
ggplot2 object.
}
\description{
Input distance matrix and metadata, and manual set metadata column names.
ggplot2 show CPCoA with color and stat ellipse.
}
\details{
By default, returns beta CPCoA coordinate
The available diversity indices include the following:
\itemize{
\item{most used indices: bray_curtis, unifrac, unifrac_binary, jaccard}
\item{other used indices: manhatten, euclidean}
}
}
\examples{
# example data: OTU table, rownames is OTU_xxx, colnames is SampleID
data(beta_bray_curtis)
# example data: metadata or design, include SampleID, genotype and site
data(metadata)
# Set 2 parameters: otu table, metadata, and distance type and groupID using default "bray_curtis" and "genotype"
beta_cpcoa_dis(beta_bray_curtis, metadata)
# Set 3 parameters: distance matrix, metadata, and groupID as "site"
beta_cpcoa_dis(beta_bray_curtis, metadata, "site")
# Set 5 parameters: distance matrix, metadata, and groupID as using "genotype", open stat elipse and sample label
beta_cpcoa_dis(beta_bray_curtis, metadata, groupID = "genotype", ellipse = T, label = T)
}
\references{
Zhang, J., Zhang, N., Liu, Y.X., Zhang, X., Hu, B., Qin, Y., Xu, H., Wang, H., Guo, X., Qian, J., et al. (2018).
Root microbiota shift in rice correlates with resident time in the field and developmental stage.
Sci China Life Sci 61, DOI: \url{https://doi.org/10.1007/s11427-018-9284-4}
}
\seealso{
beta_cpcoa_dis
}
\author{
Contact: Yong-Xin Liu \email{metagenome@126.com}
}
