% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_pcoa.R
\name{beta_pcoa}
\alias{beta_pcoa}
\title{Plotting beta diversity scatter plot}
\usage{
beta_pcoa(
  dis_mat,
  metadata,
  groupID = "genotype",
  ellipse = T,
  label = F,
  PCo = 12
)
}
\arguments{
\item{dis_mat}{distance matrix, typical output of usearch -beta_div,}

\item{metadata}{matrix or dataframe, including sampleID and groupID;}

\item{groupID}{column name for groupID.}

\item{ellipse}{stat ellipse, T or F.}

\item{label}{sample name showing, T or F.}

\item{PCo}{principle coordinate used, default 12, alternative 13, or 23.}
}
\value{
ggplot2 object.
}
\description{
Input alpha index and metadata, and manual set alpha index and metadata column names.
ggplot2 show PCoA with color and stat ellipse.
}
\details{
By default, returns beta PCoA coordinate
The available diversity indices include the following:
\itemize{
\item{most used indices: bray_curtis, unifrac}
\item{other used indices: unifrac_binary, euclidean, manhatten}
}
}
\examples{
# Set 3 parameters: dis_mat, metadata and groupID
beta_pcoa(dis_mat = beta_bray_curtis, metadata, "genotype")
# Set 2 parameters: dis_mat, metadata, and groupID as default "genotype"
beta_pcoa(dis_mat = beta_bray_curtis, metadata)
# Set 6 parameters: dis_mat, metadata, and groupID as using "site",
beta_pcoa(dis_mat = beta_unifrac, metadata, groupID = "site", ellipse = F, label = T, PCo = 13)
}
\references{
Zhang, J., Zhang, N., Liu, Y.X., Zhang, X., Hu, B., Qin, Y., Xu, H., Wang, H., Guo, X., Qian, J., et al. (2018).
Root microbiota shift in rice correlates with resident time in the field and developmental stage.
Sci China Life Sci 61, DOI: \url{https://doi.org/10.1007/s11427-018-9284-4}
}
\seealso{
beta_cpcoa
}
\author{
Contact: Yong-Xin Liu \email{metagenome@126.com}
}
