% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_pcoa_stat.R
\name{beta_pcoa_stat}
\alias{beta_pcoa_stat}
\title{Plotting beta diversity scatter plot}
\usage{
beta_pcoa_stat(
  dis_mat,
  metadata,
  groupID = "genotype",
  pairwise = TRUE,
  pairwise_list = "doc/compare.txt"
)
}
\arguments{
\item{dis_mat}{distance matrix, typical output of usearch -beta_div,}

\item{metadata}{matrix or dataframe, including sampleID and groupID;}

\item{groupID}{column name for groupID.}

\item{pairwise}{group pairwise compare, only groups less than 5, T or F.}

\item{pairwise_list}{a file include pairwise list.}
}
\value{
ggplot2 object.
}
\description{
Input alpha index and metadata, and manual set alpha index and metadata column names.
ggplot2 show PCoA with color and stat ellipse.
}
\examples{
# Set 3 parameters: dis_mat, metadata and groupID
beta_pcoa_stat(dis_mat = beta_bray_curtis, metadata, "genotype")
# Set 5 parameters: dis_mat, metadata, and groupID, pairwise FLASE, can set pairwise_list in file
beta_pcoa_stat(dis_mat, metadata, groupID = "genotype", pairwise = TRUE, pairwise_list = "doc/compare.txt")
}
\references{
Zhang, J., Zhang, N., Liu, Y.X., Zhang, X., Hu, B., Qin, Y., Xu, H., Wang, H., Guo, X., Qian, J., et al. (2018).
Root microbiota shift in rice correlates with resident time in the field and developmental stage.
Sci China Life Sci 61, DOI: \url{https://doi.org/10.1007/s11427-018-9284-4}
}
\seealso{
beta_cpcoa
}
\author{
Contact: Yong-Xin Liu \email{metagenome@126.com}
}
