% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Difference comparison}
\usage{
compare(
  data = otutab,
  metadata = metadata,
  group = "genotype",
  compare_pair = "KO-WT",
  method = "wilcox",
  RA = 0.01,
  pvalue = 0.05,
  fdr = 0.1
)
}
\arguments{
\item{data}{data matrix}

\item{metadata}{matrix or dataframe, including sampleID and groupID;}

\item{group}{column name for groupID.}

\item{compare_pair}{paired groups linked with dash.}

\item{method}{method of test.}

\item{RA}{threshold for relative abundance.}

\item{pvalue}{threshold of pvalue.}

\item{fdr}{threshold of fdr.}
}
\value{
Difference comparison table.
}
\description{
Input feature table, group info and comparing pair
Features with relative abundance, p-value and FDR for pair.
}
\details{
By default, return table
The available test methods include the following:
\itemize{
\item{most used methods: wilcoxon test, t.test}
\item{other used indices: edgeR, DESeq2}
}
}
\examples{
# Set 4 parameters
compare(data, metadata, group, compare, method)
# Set 4 parameters
compare(data, metadata, group, compare, method, RA, pvalue, fdr)
}
\references{
Jingying Zhang, Yong-Xin Liu, et. al. (2019).
NRT1.1B is associated with root microbiota composition and nitrogen use in field-grown rice.
Nature Biotechnology 37, 676-684, DOI: \url{https://doi.org/10.1038/s41587-019-0104-4}
}
\seealso{
compare_vocalno compare_heatmap compare_manhattan
}
\author{
Contact: Yong-Xin Liu \email{metagenome@126.com}
}
