% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format2maptree.R
\name{format2maptree}
\alias{format2maptree}
\title{Format feature table and taxonomy into maptree data}
\usage{
format2maptree(otu = otutab, tax = taxonomy, N = 200)
}
\arguments{
\item{otu}{OTU/ASV table;}

\item{tax}{taxonomy annotation, include feature ID and 7 levels;}

\item{N}{Top N features to show, default 200;}
}
\value{
ggplot2 object.
}
\description{
Input taxonomy composition, and feature table. Then select top N high abundance features.Finally, return a maptree(ggraph) object.
}
\details{
By default, returns top 8 taxonomy and group mean stackplot
The available style include the following:
\itemize{
\item{group: group mean circlize}
\item{sample: each sample circlize}
}
}
\examples{
# Input feature table, taxonomy and Top N features, and format into mapdata
mapdata = format2maptree(otutab, taxonomy, 200)
# Add mean abundance size and phylum color for maptree
mapadd = tax_maptree(mapdata)
# Saving and plotting maptree
(p = mapadd[[1]])
}
\references{
Zhang, J., Zhang, N., Liu, Y.X., Zhang, X., Hu, B., Qin, Y., Xu, H., Wang, H., Guo, X., Qian, J., et al. (2018).
Root microbiota shift in rice correlates with resident time in the field and developmental stage.
Sci China Life Sci 61, DOI: \url{https://doi.org/10.1007/s11427-018-9284-4}
}
\seealso{
tax_maptree
}
\author{
Contact: Yong-Xin Liu \email{metagenome@126.com}
}
