% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_circlize.R
\name{tax_circlize}
\alias{tax_circlize}
\title{Plotting circlize of taxonomy for groups or samples}
\usage{
tax_circlize(tax_sum, metadata, topN = 5, groupID = "genotype")
}
\arguments{
\item{tax_sum}{composition matrix, like OTU table and rowname is taxonomy, typical output of usearch -sintax_summary;}

\item{metadata}{matrix or dataframe, including sampleID and groupID;}

\item{topN}{Top N taxonomy to show, default 8, alternative 4, 6, 10 ...;}

\item{groupID}{column name for groupID;}

\item{style}{group or sample, default group}

\item{sorted}{Legend sorted type, default abundance, alternative alphabet}
}
\value{
ggplot2 object.
}
\description{
Input taxonomy composition, and metadata (SampleID and groupID). Then select top N high abundance taxonomy and group other low abundance. When Select samples can draw sample composition by facet groups. If used group can show mean of each group. Finally, return a ggplot2 object.
}
\details{
By default, returns top 8 taxonomy and group mean stackplot
The available style include the following:
\itemize{
\item{group: group mean circlize}
\item{sample: each sample circlize}
}
}
\examples{
# example data: feature table, rownames is OTU/taxonomy, colnames is SampleID
data(tax_phylum)
# example data: metadata or design, include SampleID, genotype and site
data(metadata)
# Set 4 parameters: set top 5 taxonomy, group by "genotype"
tax_circlize(tax_sum = tax_phylum, metadata, topN = 5, groupID = "genotype")
}
\references{
Zhang, J., Zhang, N., Liu, Y.X., Zhang, X., Hu, B., Qin, Y., Xu, H., Wang, H., Guo, X., Qian, J., et al. (2018).
Root microbiota shift in rice correlates with resident time in the field and developmental stage.
Sci China Life Sci 61, DOI: \url{https://doi.org/10.1007/s11427-018-9284-4}
}
\seealso{
tax_circlize
}
\author{
Contact: Yong-Xin Liu \email{metagenome@126.com}
}
