% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_stackplot.R
\name{tax_stackplot}
\alias{tax_stackplot}
\title{Plotting stackplot of taxonomy for groups or samples}
\usage{
tax_stackplot(
  tax_sum,
  metadata,
  topN = 8,
  groupID = "genotype",
  style = "group",
  sorted = "abundance"
)
}
\arguments{
\item{tax_sum}{composition matrix, like OTU table and rowname is taxonomy, typical output of usearch -sintax_summary;}

\item{metadata}{matrix or dataframe, including sampleID and groupID;}

\item{topN}{Top N taxonomy to show, default 8, alternative 4, 6, 10 ...;}

\item{groupID}{column name for groupID;}

\item{style}{group or sample, default group}

\item{sorted}{Legend sorted type, default abundance, alternative alphabet}
}
\value{
ggplot2 object.
}
\description{
Input taxonomy composition, and metadata (SampleID and groupID). Then select top N high abundance taxonomy and group other low abundance. When Select samples can draw sample composition by facet groups. If used group can show mean of each group. Finally, return a ggplot2 object.
}
\details{
By default, returns top 8 taxonomy and group mean stackplot
The available style include the following:
\itemize{
\item{group: group mean stackplot}
\item{sample: each sample stackplot and facet by group}
}
}
\examples{
# example data: OTU table, rownames is OTU_xxx, colnames is SampleID
data(tax_phylum)
# example data: metadata or design, include SampleID, genotype and site
data(metadata)
# Set 2 parameters: tax_sum table, metadata, default include top 8 taxonomy, groupID is genotype, show group mean and sorted by abundance
tax_stackplot(tax_sum = tax_phylum, metadata)
# Set 4 parameters: set top10 taxonomy, group by "site", and group composition
tax_stackplot(tax_sum = tax_phylum, metadata, topN = 10, groupID = "site", style = "group", sorted = "abundance")
# Set 4 parameters: set top7 taxonomy, group by "genotype", and sample composition sorted by alphabet
tax_stackplot(tax_sum = tax_phylum, metadata, topN = 10, groupID = "genotype", style = "sample", sorted = "alphabet")
}
\references{
Zhang, J., Zhang, N., Liu, Y.X., Zhang, X., Hu, B., Qin, Y., Xu, H., Wang, H., Guo, X., Qian, J., et al. (2018).
Root microbiota shift in rice correlates with resident time in the field and developmental stage.
Sci China Life Sci 61, DOI: \url{https://doi.org/10.1007/s11427-018-9284-4}
}
\seealso{
tax_stackplot
}
\author{
Contact: Yong-Xin Liu \email{metagenome@126.com}
}
